requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cu117
# torch==1.13.1
# torchvision==0.14.1
packaging
sentencepiece
peft==0.4.0
einops==0.4.1
# fastapi==0.100.1
gradio==3.39.0
markdown2==2.4.10
numpy==1.23.1
# openai==0.27.8
opencv_python==4.8.0.74
Pillow==9.4.0
pycocotools==2.0.6
ray==2.6.1
Requests==2.31.0
shortuuid==1.0.11
tqdm==4.64.1
transformers==4.31.0
# uvicorn==0.23.2
scipy~=1.11.2
bitsandbytes==0.41.1

# Lisa training
deepspeed~=0.13.1

timm~=1.0.9

# others
# imgaug~=0.4.0